add googlePlus and whatsApp sharing
